0-hash_table_create.c

+++ 0x1A-hash_tables/0-hash_table_create.c
@@ -8,7 +8,7 @@
 hash_table_t *hash_table_create(unsigned long int size)
 {
 	hash_table_t *new;
-	unsigned int i;
+	unsigned long int i;
 
 	new = malloc(sizeof(hash_table_t));
 	if (!new)
